refactor: use the same component to render the accessories and modify them
